--- a/skimage/morphology/setup.py
+++ b/skimage/morphology/setup.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python
 
+import logging
 import os
 from skimage._build import cython
+
+import pythran
 
 base_path = os.path.abspath(os.path.dirname(__file__))
 
@@ -13,7 +16,6 @@ def configuration(parent_package='', top_path=None):
 
     cython(['_skeletonize_cy.pyx',
             '_convex_hull.pyx',
-            '_grayreconstruct.pyx',
             '_extrema_cy.pyx'], working_path=base_path)
     # _skeletonize_3d uses c++, so it must be cythonized separately
     cython(['_skeletonize_3d_cy.pyx.in'], working_path=base_path)
@@ -24,8 +26,6 @@ def configuration(parent_package='', top_path=None):
     config.add_extension('_skeletonize_cy', sources=['_skeletonize_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_convex_hull', sources=['_convex_hull.c'],
-                         include_dirs=[get_numpy_include_dirs()])
-    config.add_extension('_grayreconstruct', sources=['_grayreconstruct.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_max_tree', sources=['_max_tree.c'],
                          include_dirs=[get_numpy_include_dirs()])
@@ -40,6 +40,15 @@ def configuration(parent_package='', top_path=None):
     # add precomputed footprint decomposition data
     config.add_data_files('ball_decompositions.npy',
                           'disk_decompositions.npy')
+
+    # pythran submodules
+    pythran.config.logger.setLevel(logging.INFO)
+    ext = pythran.dist.PythranExtension(
+        'skimage.morphology._grayreconstruct',
+        sources=["skimage/morphology/_grayreconstruct_pythran.py"],
+        config=['compiler.blas=none'])
+    config.ext_modules.append(ext)
+
     return config