… skimage/morphology/_grayreconstruct.pyx