@pep8speaks 2022-01-12
Hello @serge-sans-paille! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`skimage/morphology/_grayreconstruct_pythran.py`](https://github.com/scikit-image/scikit-image/blob/a5598f133b0ecf8ec8c2aedaca640b00ba9c7d6a/skimage/morphology/_grayreconstruct_pythran.py):

> [Line 2:80](https://github.com/scikit-image/scikit-image/blob/a5598f133b0ecf8ec8c2aedaca640b00ba9c7d6a/skimage/morphology/_grayreconstruct_pythran.py#L2): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (101 > 79 characters)
> [Line 13:1](https://github.com/scikit-image/scikit-image/blob/a5598f133b0ecf8ec8c2aedaca640b00ba9c7d6a/skimage/morphology/_grayreconstruct_pythran.py#L13): [E265](https://duckduckgo.com/?q=pep8%20E265) block comment should start with '# '
> [Line 16:1](https://github.com/scikit-image/scikit-image/blob/a5598f133b0ecf8ec8c2aedaca640b00ba9c7d6a/skimage/morphology/_grayreconstruct_pythran.py#L16): [E302](https://duckduckgo.com/?q=pep8%20E302) expected 2 blank lines, found 1

* In the file [`skimage/morphology/setup.py`](https://github.com/scikit-image/scikit-image/blob/a5598f133b0ecf8ec8c2aedaca640b00ba9c7d6a/skimage/morphology/setup.py):

> [Line 5:15](https://github.com/scikit-image/scikit-image/blob/a5598f133b0ecf8ec8c2aedaca640b00ba9c7d6a/skimage/morphology/setup.py#L5): [E401](https://duckduckgo.com/?q=pep8%20E401) multiple imports on one line