@hmaarrfk 2022-01-12
I might have missed a conversation, but what is the advantage of this particular implementation? it looks like a near copy paste of the cython code?

I thought pythran promised me not to write fortran code anymore :D