@serge-sans-paille 2022-01-12
@hmaarrfk I'd say that the advantage is that there's no cython-like magic
```
cdef cnp.uint32_t *ranks = <cnp.uint32_t *>(aranks.data)
```

But I need to check if there's a performance impact :-)