@grlee77 2022-01-12
Thanks @serge-sans-paille, I will take a look here soon.

Just so we don't end up duplicating effort, I wanted to let you know that I started looking at Pythran for image moment calculations (`skimage.measure.moments`). The purpose in that case was more to have a faster code path with lower memory footprint than the current code based on `numpy.dot`. To win on performance in some cases requires using the OpenMP functionality of pythran, though, since NumPy's dot may use multi-threaded OpenBLAS or MKL.