@grlee77 2022-01-14
> note: the benchmark fail to run, but the error looks unrelated to me...

yeah it looks like it was trying to build NumPy for some reason. I just updated the ASV configuration to see if that helps and will start the benchmarks again