@grlee77 2022-01-15
I updated the problem size to something taking closer to 0.1-1 second. Timings are the same to within 1% after the changes here.

The easiest way I found to benchmark against older commits where the benchmark wasn't yet present is to just do:

1.) run the benchmark in the dev branch using `asv run --python same -b Reconstruction` where the "-b Reconstruction" is just to save time by only running the benchmarks corresponding to the new class introduced here.
2.) make a copy of the file with the new benchmarks
3.) checkout main
4.) repeat step 1 (it should now find the file with the benchmark copy from step 2)


### Result

On main:
```
[100.00%] ··· bench_morph.Reconstruction.time_reconstruction                  ok
[100.00%] ··· ====== ============
               ndim              
              ------ ------------
                1     47.3±0.6ms 
                2      442±3ms   
                3      662±8ms   
              ====== ============

```

With this PR:
```
[100.00%] ··· ...orphology.Reconstruction.time_reconstruction                 ok
[100.00%] ··· ====== ============
               ndim              
              ------ ------------
                1     47.2±0.4ms 
                2      448±8ms   
                3      659±8ms   
              ====== ============
```