@serge-sans-paille 2022-09-28
@grlee77 I've rebased the PR and solved the merge conflicts, let see how this one performs wrt. benchmarks.