@serge-sans-paille 2022-09-29
@grlee77 even if asv doesn't detect a performance regression, I'm curious about the actual numbers. Where can I collect them?