@lagru 2022-09-30
I think you figured it out already, seeing that you added the `run-benchmarks` label. Have a look at the log and artifacts in https://github.com/scikit-image/scikit-image/actions/runs/3157167004. Does that contain your numbers?