@serge-sans-paille 2022-09-30
Yes! for `GrayReconstruction.time_reconstruction`

before
```
============== =============== ===========
    shape           dtype                 
-------------- --------------- -----------
   (10, 10)      numpy.uint8    235±0.3μs     
   (10, 10)     numpy.float32   234±0.4μs 
   (10, 10)     numpy.float64   235±0.3μs 
   (64, 64)      numpy.uint8     720±1μs       
   (64, 64)     numpy.float32    716±7μs  
   (64, 64)     numpy.float64    726±4μs  
 (1200, 1200)    numpy.uint8    275±0.7ms     
 (1200, 1200)   numpy.float32    265±1ms  
 (1200, 1200)   numpy.float64    275±2ms  
 (96, 96, 96)    numpy.uint8     162±2ms       
 (96, 96, 96)   numpy.float32   157±0.7ms 
 (96, 96, 96)   numpy.float64    161±1ms  
============== =============== ===========
```

after

```
============== =============== ===========
    shape           dtype                 
-------------- --------------- -----------
   (10, 10)      numpy.uint8    236±0.6μs     
   (10, 10)     numpy.float32   235±0.6μs 
   (10, 10)     numpy.float64   235±0.4μs 
   (64, 64)      numpy.uint8     715±4μs       
   (64, 64)     numpy.float32    718±4μs  
   (64, 64)     numpy.float64    731±6μs  
 (1200, 1200)    numpy.uint8     272±1ms       
 (1200, 1200)   numpy.float32    263±1ms  
 (1200, 1200)   numpy.float64    272±2ms  
 (96, 96, 96)    numpy.uint8    159±0.6ms     
 (96, 96, 96)   numpy.float32   155±0.7ms 
 (96, 96, 96)   numpy.float64    159±1ms  
============== =============== ===========
```