@lagru 2022-09-30
Wow, I completely missed that we [support Pythran in scikit-image since 0.19](https://scikit-image.org/docs/0.19.x/release_notes.html#announcement-scikit-image-0-19-0) and mistook this for a "demo". Is this advantage described in https://github.com/scikit-image/scikit-image/pull/6182#issuecomment-1011050037 still the reason for this PR? In that case I would be happy to look into this and familiarise myself with Pythran a little more.