@grlee77 2022-09-30
Sorry about not replying sooner. Running the benchmarks locally should be more reliable for detecting small changes since the same hardware will be used. There is a little documentation about commands to compare two commits in the contributing guide [here](https://github.com/scikit-image/scikit-image/blob/main/CONTRIBUTING.rst#benchmarks).

On CI, we can't guarantee we get exactly the same hardware/conditions across the two runs, so it is not ideal for detecting small differences (especially longitudinally over time). Jaime wrote a [blog post](https://labs.quansight.org/blog/2021/08/github-actions-benchmarks) indicating the typical level of spread in results over repeated runs. IIRC the conclusion is things that change performance more than 20% or so can be pretty reliably detected.